requirements.txt
hydra-core>=1.3
omegaconf>=2.3
optuna>=3.0
torch
torchvision
tqdm